third_party/WebKit/LayoutTests/fast/dom/Geolocation/notimer-after-unload-expected.txt

main frame - has 1 onunload handler(s)
CONSOLE ERROR: Blocked alert('unload-called') during unload.
You should have seen one unload alert appear.
